std: simplify the random feature
